backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Pillow==9.5.0
ExifRead==3.0.0
